+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Official Releases:
 version **0.2.3** (?? ?? ??)
 *******************************
 
+* added the ``save_page_source`` action.
 * added a parameter to ``take_screnshot`` to include a timestamp in the file 
   name
 * expose the SSTTestCase class which is used internally to create test cases